refactor: update Button imports and fix prop casing
- Update all Button imports to use new ui/ directory path
- Fix autofocus → autoFocus prop casing (React convention)
- Update Footer component to use Button.tsx
