+++ src/components/ForkMonitor.tsx
@@ -11,7 +11,7 @@ export const ForkMonitor: React.FC<ForkMonitorProps> = ({ animated = true }) => 
   return (
     <ForkDataProvider>
       <ForkMockProvider>
-        <ForkDisplay animated={animated} />
+        <ForkDisplay />
       </ForkMockProvider>
     </ForkDataProvider>
   )
