+++ src/layouts/LearnLayout.astro
@@ -3,7 +3,7 @@ import Layout from "../layouts/Layout.astro";
 import SocialLinks from "../components/SocialLinks.astro";
 import Footer from "../components/Footer.astro";
 import LearnNavigation from "../components/LearnNavigation.tsx";
-import Button from "../components/Button.tsx";
+import Button from "../components/ui/Button.tsx";
 import Pointer from "../components/Pointer.tsx";
 import PageHeading from "../components/PageHeading.astro";
 import { withBase } from "../lib/utils";
@@ -38,7 +38,7 @@ const currentPath = Astro.url.pathname
 		{/* Top Section: Navigation */}
 		<div class="flex justify-between items-start p-8">
 			{/* Back Button */}
-			<Button id="back-button" variant="link" href={withBase('/?intro=false')} autofocus>
+			<Button id="back-button" variant="link" href={withBase('/?intro=false')} autoFocus>
 				<Pointer direction="left" animated="auto" />
 				BACK TO HOME
 			</Button>
