+++ src/pages/team.astro
@@ -3,7 +3,7 @@ import Layout from "../layouts/Layout.astro";
 import SocialLinks from "../components/SocialLinks.astro";
 import TeamCard from "../components/TeamCard.astro";
 import Footer from "../components/Footer.astro";
-import Button from "../components/Button.tsx";
+import Button from "../components/ui/Button.tsx";
 import Pointer from "../components/Pointer.tsx";
 import PageHeading from "../components/PageHeading.astro";
 import SectionHeading from "../components/SectionHeading.astro";
@@ -17,7 +17,7 @@ import { withBase } from "../lib/utils";
     <!-- Top Section: Navigation -->
     <div class="flex justify-between items-start p-8">
       <!-- Back Button -->
-      <Button id="back-button" variant="link" size="lg" href={withBase('/?intro=false')} autofocus>
+      <Button id="back-button" variant="link" size="lg" href={withBase('/?intro=false')} autoFocus>
         <Pointer animated="auto" direction="left" />
         BACK TO HOME
       </Button>
